Integrate initial client location load and google maps functionality to dashboard page
