+++ client/src/pages/LandingPage.jsx
@@ -4,7 +4,6 @@ import { Link } from 'react-router-dom'
 import bar from '../assets/images/bar.jpg';
 import LoginModal from '../components/LoginModal';
 import RegisterModal from '../components/RegisterModal';
-import Map from '../components/Maps';
 
 const LandingPage = () => {
   const [ showLogin, setShowLogin ] = useState(false);
@@ -13,7 +12,6 @@ const LandingPage = () => {
 
   return (
     <>
-      <Map></Map>
       <div className='top'>
         <Header
           setShowLogin={setShowLogin}
