+++ server/db.js
@@ -1,5 +1,5 @@
 //CONNECT  SERVER
-
+require('dotenv').config()
 const { Sequelize } = require('sequelize');
 // console.log(process.env.URI, 'URI STRING');
